# Conflicts:
#	build.gradle
#	src/main/java/com/example/BaeGongPaServer/Auth/AuthProvider.java
#	src/main/java/com/example/BaeGongPaServer/Component/ApiResponse.java
#	src/main/java/com/example/BaeGongPaServer/Controller/AuthController.java
#	src/main/java/com/example/BaeGongPaServer/DAO/AuthUserDAO.java
#	src/main/java/com/example/BaeGongPaServer/DTO/AuthDTO.java
#	src/main/java/com/example/BaeGongPaServer/DTO/MemInfoDto.java
#	src/main/java/com/example/BaeGongPaServer/DTO/RoomInfoDto.java
#	src/main/java/com/example/BaeGongPaServer/Domain/MemInfo.java
#	src/main/java/com/example/BaeGongPaServer/Domain/MenuBoxInfo.java
#	src/main/java/com/example/BaeGongPaServer/Domain/MenuInfo.java
#	src/main/java/com/example/BaeGongPaServer/Domain/OrderInfo.java
#	src/main/java/com/example/BaeGongPaServer/Domain/RoomInfo.java
#	src/main/java/com/example/BaeGongPaServer/Repository/MemInfoRepository.java
#	src/main/java/com/example/BaeGongPaServer/Repository/OrderInfoRepository.java
#	src/main/java/com/example/BaeGongPaServer/Repository/RoomInfoRepository.java
#	src/main/java/com/example/BaeGongPaServer/Service/AuthService.java
#	src/main/java/com/example/BaeGongPaServer/Service/MemberInfoService.java
#	src/main/resources/application.properties

+++ src/main/java/com/example/BaeGongPaServer/DAO/AuthUserDAO.java
@@ -13,7 +13,7 @@ import java.util.Collection;
 @Setter
 public class AuthUserDAO implements UserDetails {
     private long idx;
-    private Integer memNo;
+    private Long memNo;
     private String memNick;
     private String memId;
     private String memPwd;
